rename contact-summary to contact-plaque, reduce to followers and flaggers
