+++ src/com/index.js
@@ -235,7 +235,7 @@ exports.messageSummary = require('./message-summary')
 exports.messageAttachments = require('./message-attachments')
 exports.messageStats = require('./message-stats')
 exports.contactFeed = require('./contact-feed')
-exports.contactSummary = require('./contact-summary')
+exports.contactPlaque = require('./contact-plaque')
 exports.contactListing = require('./contact-listing')
 exports.programFeed = require('./program-feed')
 exports.programSummary = require('./program-summary')
